heartbeat: improve how we process
We need to make sure that heartbeat is ignored during connecting and
that half time flag is reset on heartbeat update. As well as we need
to make sure that last updated incoming and outgoing time is atomic.
We need atomic update to last incoming or outgoing heartbeat to make
sure all threads see the change.

With the change we added number of tests around heartbeat handling
which required adding now method to easly mock time in millis used
in the code.

Bug-Url: https://bugzilla.redhat.com/1710818
Change-Id: I1d1399c661270fee2a9c688f7d9ba69fa006a16b
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>