Java 11 migration
Due to xmvn clash with java 11 maven-toolchain-plugin is used as a
workaround to allow java 1.8 as xmvn runtime and java 11 as runtime for
build, compilation and basically a runtime for mvn plugins.
https://bugzilla.redhat.com/show_bug.cgi?id=1761381

One drawback of using toolchain plugin is a necessity to enable bootstrap
option ie. %mvn_build -b  because toolchain plugin for some reason is
not resolved even though installed as rh-maven35-toolchain-plugin rpm.

Toochain is enabled for el7 only as maven profile. Once xmvn package is
upgraded it is supposed to be removed.

Change-Id: Ie11b8c27f33100468050037e3572dff523cec0b5
Bug-Url: https://bugzilla.redhat.com/1734729
Signed-off-by: Artur Socha <asocha@redhat.com>