HeartbeatTestCase fixed
This patch re-enables heartbeat unit tests  and fixes policy setup.
Additionally, in order to avoid getting out of sync with remaining
codebase, new maven surefire profile is created. This profile is
supposed to run all the unit tests (including those long-running that
are currently ignored)
With time, more unit tests marked as @Ignored will be migrated to this
profile so that CI would be properly utilized.

Change-Id: I5a884a8431a5fe929750526159f7c24861ed2572
Signed-off-by: Artur Socha <asocha@redhat.com>