Fail to re-connect with expired ssl certificates
Before this change, the ssl handshake has been initiated after a client
established a plain connection. Unfortunately, SSL certificates validation
was happening too late and the connection was being scheduled to close
and was entering into the infinite loop (connect -> close -> connect -> close
etc).
That caused an issue on the ovirt engine's side because it was not able
to determine when to move the host from the 'connecting' state to
'non_responsive'. The failure manifested itself in the error response
message.
This patch allows to detect ssh certificates validation issue much
earlier (just after establishing the connection) so that a regular
re-try mechanism is used and when it reaches max retries, then it notifies
the host about it and the host becomes 'non_responsive' (as expected).

While the host is 'non_responsive' then it is possible to enroll new
valid certificate.

Signed-off-by: Artur Socha <asocha@redhat.com>