Disconnect the session after SSL handshake issue
When the handshake check that is added in PR #17 fails, then a
ClientConnectionException is thrown, and passed up to the client in
ovirt-engine.
But as the connection is in some invalid state, but not disconnected, it
for example causes there are no heartbeats anymore.

This fix also closes the connection when there is some
ClientConnectionException in the postConnect stage.

This should fix #27

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>